{{cookiecutter.project_name}}/packaged-requirements.txt
pyspark-shared-libs-8451==1.0.19